+++ example/lib/fields_class_example.g.dart
@@ -58,7 +58,7 @@ mixin _$Mother {
 }
 
 class _MotherFields {
-  // ignore: unused_element
+  // ignore: unused_element, unused_element_parameter
   const _MotherFields([this._path = '']);
 
   final String _path;