+++ example/lib/typedef_example.dart
@@ -1,17 +1,19 @@
-import 'package:example/basic_example.dart' as be;
+import 'package:example/basic_example.dart';
+import 'package:json_annotation/json_annotation.dart';
 import 'package:mek_data_class/mek_data_class.dart';
 
 part 'typedef_example.g.dart';
 
 typedef JsonMap<T> = Map<String, T>;
-typedef NullableJsonMap<T> = Map<String?, T>;
+typedef NullableJsonMap<T> = Map<String, T>;
 
 @DataClass(changeable: true, copyable: true)
+@JsonSerializable(createToJson: false)
 class Extra with _$Extra {
   final JsonMap<int>? jsonMap;
   final NullableJsonMap<double?> nullableJsonMap;
-  final be.Product product;
-  final be.Product product2;
+  final Product product;
+  final Product product2;
 
   const Extra({
     required this.jsonMap,
@@ -19,4 +21,6 @@ class Extra with _$Extra {
     required this.product,
     required this.product2,
   });
+
+  factory Extra.fromJson(Map<String, dynamic> map) => _$ExtraFromJson(map);
 }